Moved options objects to new package io.vertx.redis.op
