--- src/main/java/io/vertx/redis/InsertOptions.java
@@ -1,9 +1,0 @@
-package io.vertx.redis;
-
-/**
- * @author <a href="mailto:marko.strukelj@gmail.com">Marko Strukelj</a>
- */
-public enum InsertOptions {
-  BEFORE,
-  AFTER
-}
