+++ src/main/java/io/vertx/redis/RedisService.java
@@ -9,6 +9,19 @@ import io.vertx.core.Vertx;
 import io.vertx.core.json.JsonArray;
 import io.vertx.core.json.JsonObject;
 import io.vertx.redis.impl.RedisServiceImpl;
+import io.vertx.redis.op.AggregateOptions;
+import io.vertx.redis.op.BitOperation;
+import io.vertx.redis.op.InsertOptions;
+import io.vertx.redis.op.KillFilter;
+import io.vertx.redis.op.LimitOptions;
+import io.vertx.redis.op.MigrateOptions;
+import io.vertx.redis.op.ObjectCmd;
+import io.vertx.redis.op.RangeLimitOptions;
+import io.vertx.redis.op.RangeOptions;
+import io.vertx.redis.op.ScanOptions;
+import io.vertx.redis.op.SetOptions;
+import io.vertx.redis.op.ShutdownOptions;
+import io.vertx.redis.op.SortOptions;
 import io.vertx.serviceproxy.ProxyHelper;
 
 import java.util.List;
