+++ src/main/java/io/vertx/redis/impl/RedisServiceImpl.java
@@ -5,20 +5,20 @@ import io.vertx.core.Handler;
 import io.vertx.core.Vertx;
 import io.vertx.core.json.JsonArray;
 import io.vertx.core.json.JsonObject;
-import io.vertx.redis.AggregateOptions;
-import io.vertx.redis.BitOperation;
-import io.vertx.redis.InsertOptions;
-import io.vertx.redis.KillFilter;
-import io.vertx.redis.LimitOptions;
-import io.vertx.redis.MigrateOptions;
-import io.vertx.redis.ObjectCmd;
-import io.vertx.redis.RangeLimitOptions;
-import io.vertx.redis.RangeOptions;
+import io.vertx.redis.op.AggregateOptions;
+import io.vertx.redis.op.BitOperation;
+import io.vertx.redis.op.InsertOptions;
+import io.vertx.redis.op.KillFilter;
+import io.vertx.redis.op.LimitOptions;
+import io.vertx.redis.op.MigrateOptions;
+import io.vertx.redis.op.ObjectCmd;
+import io.vertx.redis.op.RangeLimitOptions;
+import io.vertx.redis.op.RangeOptions;
 import io.vertx.redis.RedisEncoding;
-import io.vertx.redis.ScanOptions;
-import io.vertx.redis.SetOptions;
-import io.vertx.redis.ShutdownOptions;
-import io.vertx.redis.SortOptions;
+import io.vertx.redis.op.ScanOptions;
+import io.vertx.redis.op.SetOptions;
+import io.vertx.redis.op.ShutdownOptions;
+import io.vertx.redis.op.SortOptions;
 
 import java.util.Collection;
 import java.util.Collections;
