+++ src/main/java/io/vertx/redis/op/AggregateOptions.java
@@ -1,19 +1,22 @@
-package io.vertx.redis;
+package io.vertx.redis.op;
 
 import io.vertx.core.json.JsonArray;
 
 /**
  * @author <a href="mailto:marko.strukelj@gmail.com">Marko Strukelj</a>
  */
-public enum RangeOptions {
+public enum AggregateOptions {
   NONE,
-  WITHSCORES;
+  SUM,
+  MIN,
+  MAX;
 
   public JsonArray toJsonArray() {
     JsonArray result = new JsonArray();
     if (this == NONE) {
       return result;
     }
+    result.add("AGGREGATE");
     result.add(name());
     return result;
   }
