+++ src/main/java/io/vertx/redis/op/BitOperation.java
@@ -1,10 +1,11 @@
-package io.vertx.redis;
+package io.vertx.redis.op;
 
 /**
  * @author <a href="mailto:marko.strukelj@gmail.com">Marko Strukelj</a>
  */
-public enum ObjectCmd {
-  REFCOUNT,
-  ENCODING,
-  IDLETIME
+public enum BitOperation {
+  AND,
+  OR,
+  XOR,
+  NOT
 }
