+++ src/main/java/io/vertx/redis/op/InsertOptions.java
@@ -1,11 +1,9 @@
-package io.vertx.redis;
+package io.vertx.redis.op;
 
 /**
  * @author <a href="mailto:marko.strukelj@gmail.com">Marko Strukelj</a>
  */
-public enum BitOperation {
-  AND,
-  OR,
-  XOR,
-  NOT
+public enum InsertOptions {
+  BEFORE,
+  AFTER
 }
