+++ src/main/java/io/vertx/redis/op/ObjectCmd.java
@@ -1,0 +1,10 @@
+package io.vertx.redis.op;
+
+/**
+ * @author <a href="mailto:marko.strukelj@gmail.com">Marko Strukelj</a>
+ */
+public enum ObjectCmd {
+  REFCOUNT,
+  ENCODING,
+  IDLETIME
+}
