+++ src/main/java/io/vertx/redis/op/RangeOptions.java
@@ -1,22 +1,19 @@
-package io.vertx.redis;
+package io.vertx.redis.op;
 
 import io.vertx.core.json.JsonArray;
 
 /**
  * @author <a href="mailto:marko.strukelj@gmail.com">Marko Strukelj</a>
  */
-public enum AggregateOptions {
+public enum RangeOptions {
   NONE,
-  SUM,
-  MIN,
-  MAX;
+  WITHSCORES;
 
   public JsonArray toJsonArray() {
     JsonArray result = new JsonArray();
     if (this == NONE) {
       return result;
     }
-    result.add("AGGREGATE");
     result.add(name());
     return result;
   }
