+++ src/main/java/io/vertx/redis/op/ScanOptions.java
@@ -14,7 +14,7 @@
  *  You may elect to redistribute this code under either of these licenses.
  */
 
-package io.vertx.redis;
+package io.vertx.redis.op;
 
 import io.vertx.codegen.annotations.DataObject;
 import io.vertx.core.json.JsonArray;
