+++ src/main/java/io/vertx/redis/op/SortOptions.java
@@ -1,4 +1,4 @@
-package io.vertx.redis;
+package io.vertx.redis.op;
 
 import io.vertx.codegen.annotations.DataObject;
 import io.vertx.core.json.JsonArray;
